+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## [1.18] - 5/13/2024
 
 - Modified compare list
-- Add Strongs Bibles in menu 
+- Added Strongs Bibles in menu
+- Added Strongs search 
 
 ## [1.17] - 2/22/2024